seperated gesture recognizer delegate into a new class and fixed downstream breaks
